refactor: mv post mdl code to sep files, posts dir
Move interfaces, classes, and other units  in files in models/posts
and model/GeneralPosts to separate files.  Moved everything in
model/GeneralPosts to model/posts.
