+++ web-client/src/components/TopNavbar/TopNavbar.tsx
@@ -21,7 +21,7 @@ import {
   AlgFindRequestsLocation,
   MyOfferPostsLocationUrl,
   MyRequestPostsLocationUrl,
-} from 'src/modules/myRequests/constants';
+} from 'src/modules/requestAndOfferPosts/constants';
 import { AppState } from 'src/store';
 import { COLORS } from 'src/theme/colors';
 import styled from 'styled-components';
