+++ web-client/src/ducks/MyRequests/actions.ts
@@ -1,4 +1,5 @@
-import { IPost, Post } from 'src/models/posts';
+import { Post } from "src/models/Post";
+import { IPost } from "src/models/IPost";
 import { User } from 'src/models/users';
 import { createPost, observePosts } from 'src/services/posts';
