+++ web-client/src/models/posts/GeneralPost.ts
@@ -5,36 +5,17 @@ import {
   IsNumber,
   IsObject,
   IsOptional,
-  IsString,
+  IsString
 } from 'class-validator';
 import { firestore } from 'firebase';
 import { firestore as db } from 'src/firebase';
+import { Post } from "../Post";
+import { IPost } from "../IPost";
+import { PostStatus } from "../PostStatus";
+import { UnauthenticatedPost } from "../UnauthenticatedPost";
+import { IGeneralPost } from './IGeneralPost';
+import { IUserGeneral } from "../users/IUserGeneral";
 
-import { IPost, Post, PostStatus } from '../posts';
-import {
-  IUnauthenticatedPost,
-  UnauthenticatedPost,
-} from '../UnauthenticatedPost';
-
-interface IUserGeneral {
-  displayName: string;
-  displayPicture: string | null;
-  rating: number | null;
-}
-
-export interface IGeneralPost extends IUnauthenticatedPost {
-  creatorRef: string;
-  creatorSnapshot: IUserGeneral;
-  status: PostStatus;
-  streetAddress: string;
-  participants: string[];
-  rejected: string[];
-  offerCount: number;
-  rejectionCount: number;
-  firstOfferMade: Date | null;
-  firstRejectionMade: Date | null;
-  seenBy: string[];
-}
 
 export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
   constructor(
@@ -55,7 +36,7 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
     firstRejectionMade: Date | null = null,
     seenBy: string[] = [],
     createdAt?: Date,
-    updatedAt?: Date,
+    updatedAt?: Date
   ) {
     super(
       postRef,
@@ -68,7 +49,7 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
       description,
       latLng,
       createdAt,
-      updatedAt,
+      updatedAt
     );
     this._creatorSnapshotGeneral = creatorSnapshot;
     this._creatorRef = creatorRef;
@@ -251,8 +232,7 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
       {
         displayName: data.creatorSnapshot.displayName || '',
         displayPicture: data.creatorSnapshot.displayPicture,
-        rating:
-          data.creatorSnapshot.pinRatingsReceived /
+        rating: data.creatorSnapshot.pinRatingsReceived /
           data.creatorSnapshot.requestsMade,
       },
       data.title,
@@ -272,12 +252,12 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
       data.firstRejectionMade?.toDate(),
       seenBy,
       data.createdAt.toDate(),
-      data.updatedAt.toDate(),
+      data.updatedAt.toDate()
     );
   }
 
   public static fromFirestore(
-    data: firebase.firestore.DocumentData,
+    data: firebase.firestore.DocumentData
   ): GeneralPost {
     return new GeneralPost(
       (data.requestRef as firebase.firestore.DocumentReference).path,
@@ -300,7 +280,7 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
       (data.firstRejectionMade as firebase.firestore.Timestamp).toDate(),
       data.seenBy,
       (data.createdAt as firebase.firestore.Timestamp).toDate(),
-      (data.updatedAt as firebase.firestore.Timestamp).toDate(),
+      (data.updatedAt as firebase.firestore.Timestamp).toDate()
     );
   }
 
@@ -323,7 +303,7 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
       data.firstRejectionMade,
       data.seenBy,
       data.createdAt,
-      data.updatedAt,
+      data.updatedAt
     );
   }
 
@@ -354,7 +334,7 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
       data.firstRejectionMade,
       data.seenBy,
       data.createdAt,
-      data.updatedAt,
+      data.updatedAt
     );
   }
 
@@ -390,7 +370,7 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
       description: this.description,
       latLng: new firestore.GeoPoint(
         this.latLng.latitude,
-        this.latLng.longitude,
+        this.latLng.longitude
       ),
       creatorRef: db.doc(this.creatorRef),
       status: this.status,
@@ -438,11 +418,3 @@ export class GeneralPost extends UnauthenticatedPost implements IGeneralPost {
     };
   }
 }
-
-export const GeneralPostFirestoreConverter: firebase.firestore.FirestoreDataConverter<GeneralPost> = {
-  fromFirestore: (
-    data: firebase.firestore.QueryDocumentSnapshot<IGeneralPost>,
-  ): GeneralPost => GeneralPost.fromFirestore(data.data()),
-  toFirestore: (modelObject: GeneralPost): firebase.firestore.DocumentData =>
-    modelObject.toFirestore(),
-};
