+++ web-client/src/models/posts/ILatLngObject.ts
@@ -1,0 +1,4 @@
+export interface ILatLngObject {
+  latitude: number;
+  longitude: number;
+}
