+++ web-client/src/models/posts/PostStatus.ts
@@ -1,0 +1,11 @@
+
+export enum PostStatus {
+  pending = 'pending',
+  ongoing = 'ongoing',
+  completed = 'completed',
+  declined = 'declined',
+  closed = 'closed',
+  open = 'open',
+  active = 'active',
+  removed = 'removed'
+}
