+++ web-client/src/models/users/IStrippedUser.ts
@@ -1,0 +1,4 @@
+export interface IStrippedUser {
+  displayName: string;
+  displayPicture: string | null;
+}
