+++ web-client/src/models/users/IUserGeneral.ts
@@ -1,0 +1,6 @@
+
+export interface IUserGeneral {
+  displayName: string;
+  displayPicture: string | null;
+  rating: number | null;
+}
