+++ web-client/src/modules/create/containers/CreatePostContainer.tsx
@@ -7,14 +7,15 @@ import StepTracker from 'src/components/StepTracker/StepTracker';
 import { createOffer } from 'src/ducks/MyOffers/actions';
 import { createRequest } from 'src/ducks/MyRequests/actions';
 import { ProfileState } from 'src/ducks/profile/types';
-import { IPost, PostStatus } from 'src/models/posts';
+import { IPost } from "src/models/IPost";
+import { PostStatus } from "src/models/PostStatus";
 import { IUser } from 'src/models/users';
 import { IUserAddress } from 'src/models/users/privilegedInformation';
 import NewAddressModal from 'src/modules/create/components/NewAddressModal';
 import PostDetailsStep from 'src/modules/create/components/PostDetailsStep';
 import PostLocationStep from 'src/modules/create/components/PostLocationStep';
 import PostSummary from 'src/modules/create/components/PostSummaryStep';
-import { MyRequestPostsLocationUrl } from 'src/modules/myRequests/constants';
+import { MyRequestPostsLocationUrl } from 'src/modules/requestAndOfferPosts/constants';
 import AuthenticationModal from 'src/pages/modals/AuthenticationModal';
 import { AppState } from 'src/store';
 import styled from 'styled-components';
