+++ web-client/src/modules/requestAndOfferPosts/components/PostInfo.tsx
