+++ web-client/src/modules/requestAndOfferPosts/containers/MyOffersContainer.tsx
