+++ web-client/src/modules/timeline/containers/TimelineViewContainer/TimelineViewContainer.tsx
@@ -7,12 +7,12 @@ import { MyRequestsState } from 'src/ducks/MyRequests/types';
 import { ProfileState } from 'src/ducks/profile/types';
 import { getPostWithOffersAndTimelineItems } from 'src/ducks/timeline/functions';
 import { firestore as firestore2 } from 'src/firebase';
-import { Post } from 'src/models/posts';
+import { Post } from "src/models/Post";
 import { ApplicationPreference } from 'src/models/users';
 import {
   MyOfferPostsLocationUrl,
   MyRequestPostsLocationUrl,
-} from 'src/modules/myRequests/constants';
+} from 'src/modules/requestAndOfferPosts/constants';
 
 import LoadingWrapper from '../../../../components/LoadingComponent/LoadingComponent';
 import {
